fix: 抽取 type
