+++ src/maps/util/mapbox.ts
@@ -1,6 +1,7 @@
-import { Mapbox, IMapOptions } from '@antv/l7-maps';
+import { IMapConfig } from '@antv/l7';
+import { Mapbox } from '@antv/l7-maps';
 
-export const CreateMapBox = (props: IMapOptions) => {
+export const CreateMapBox = (props: IMapConfig) => {
   const { mapCinfig } = props;
   return new Mapbox({
     mapCinfig,
